+++ src/prepare-runtime.ts
@@ -1,10 +1,10 @@
 import type { Nuxt } from '@nuxt/schema'
 import type { I18nNuxtContext } from './context'
-import { VIRTUAL_NUXT_I18N_LOGGER } from './virtual-logger'
+import { i18nVirtualLoggerPlugin, RESOLVED_VIRTUAL_NUXT_I18N_LOGGER, VIRTUAL_NUXT_I18N_LOGGER } from './virtual-logger'
 import { defu } from 'defu'
-import { addPlugin, addTemplate } from '@nuxt/kit'
+import { addPlugin, addTemplate, addTypeTemplate } from '@nuxt/kit'
 import { generateTemplateNuxtI18nOptions } from './template'
-import { generateLoaderOptions, simplifyLocaleOptions } from './gen'
+import { generateI18nTypes, generateLoaderOptions, simplifyLocaleOptions } from './gen'
 import { NUXT_I18N_TEMPLATE_OPTIONS_KEY } from './constants'
 
 export function prepareRuntime(ctx: I18nNuxtContext, nuxt: Nuxt) {
@@ -47,5 +47,25 @@ export function prepareRuntime(ctx: I18nNuxtContext, nuxt: Nuxt) {
     filename: NUXT_I18N_TEMPLATE_OPTIONS_KEY,
     write: true,
     getContents: () => genTemplate(false)
+  })
+
+  nuxt.options.imports.transform ??= {}
+  nuxt.options.imports.transform.include ??= []
+  nuxt.options.imports.transform.include.push(new RegExp(`${RESOLVED_VIRTUAL_NUXT_I18N_LOGGER}$`))
+
+  nuxt.hook('vite:extendConfig', cfg => {
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+    cfg.plugins ||= []
+    // @ts-ignore NOTE: A type error occurs due to a mismatch between Vite plugins and those of Rollup
+    cfg.plugins.push(i18nVirtualLoggerPlugin(options.debug))
+  })
+
+  /**
+   * `$i18n` type narrowing based on 'legacy' or 'composition'
+   * `locales` type narrowing based on generated configuration
+   */
+  addTypeTemplate({
+    filename: 'types/i18n-plugin.d.ts',
+    getContents: () => generateI18nTypes(nuxt, ctx.userOptions)
   })
 }
